chore: adjust structure foldering
