+++ src/app/[locale]/(public)/events/[id]/page.tsx
@@ -1,8 +1,8 @@
-import { EventDetailPage } from "@/features/events";
+import { PublicEventDetailPage } from "@/features/events/pages";
 
 const EventsDetail = async (props: { params: Promise<{ id: string }> }) => {
   const params = await props.params;
-  return <EventDetailPage eventId={params.id} />;
+  return <PublicEventDetailPage eventId={params.id} />;
 };
 
 export default EventsDetail;
