--- src/features/events/admin/index.ts
@@ -1,2 +1,0 @@
-export { default as AdminEventsListPage } from "./AdminEventsListPage";
-export { default as AdminEventsCreatePage } from "./AdminEventsCreatePage";
