+++ src/features/events/hooks/useMyEvent.ts
@@ -7,19 +7,19 @@ export const useMyEvents = (page: number = 1, limit: number = 10) => {
   const [myEvents, setMyEvents] = useState<EventType[]>([]);
   const [isLoading, setIsLoading] = useState<boolean>(false);
 
-  useEffect(() => {
-    const getEvents = async () => {
-      setIsLoading(true);
-      try {
-        const res = await eventsService.getMyEvents(page, limit);
-        setMyEvents(res.data);
-      } catch (err) {
-        toast.error(err instanceof Error ? err.message : "Something went wrong.");
-      } finally {
-        setIsLoading(false);
-      }
-    };
+  const getEvents = async () => {
+    setIsLoading(true);
+    try {
+      const res = await eventsService.getMyEvents(page, limit);
+      setMyEvents(res.data);
+    } catch (err) {
+      toast.error(err instanceof Error ? err.message : "Something went wrong.");
+    } finally {
+      setIsLoading(false);
+    }
+  };
 
+  useEffect(() => {
     getEvents();
   }, [toast]);
