--- src/features/events/index.tsx
@@ -1,2 +1,0 @@
-export { default as EventListPage } from "./EventListPage";
-export { default as EventDetailPage } from "./EventDetailPage";
