+++ src/features/events/pages/index.ts
@@ -1,0 +1,5 @@
+export { default as AdminEventsListPage } from "./AdminEventsListPage";
+export { default as AdminEventsCreatePage } from "./AdminEventsCreatePage";
+export { default as PublicEventListPage } from "./PublicEventListPage";
+export { default as PublicEventDetailPage } from "./PublicEventDetailPage";
+export { default as UserEventPage } from "./UserEventPage";
