+++ src/features/profile/pages/index.ts
@@ -1,0 +1,1 @@
+export { default as UserProfilePage } from "./UserProfilePage";
